moving middleware to core and removing 'use' until we flesh out middleware (in next version)
