+++ lib/assemble.js
@@ -59,7 +59,7 @@ var Assemble = function (src, config) {
 
   this.src = config.src;
 
-  var omissions = _.omit(this.defaults, ['plugins', 'middleware']);
+  var omissions = _.omit(this.defaults, ['plugins', 'core']);
   this.options = this.config = _.merge({}, omissions, config || {}, runtimeConfig || {});
 
   // Clone the config so they can be accessed through the api
@@ -71,8 +71,6 @@ var Assemble = function (src, config) {
   this.config.plugins = _.union(Assemble.defaults.plugins, this.config.plugins);
   this.plugins = this.dovetail.resolve(this.config.plugins);
 
-  this.middleware = [];
-
   // setup the context
   this.context = this.utils.context.stub(this);
 
@@ -83,7 +81,7 @@ Assemble.utils  = require('./utils');
 Assemble.render = require('./render');
 Assemble.logger = Assemble.utils.log;
 
-var middleware = require('./middleware');
+var core = require('./core');
 Assemble.models = require('./models');
 
 // add assemble plugin events to Dovetail
@@ -102,37 +100,32 @@ Assemble.defaults = _.merge({
     relative(path.join(Assemble.paths.plugins, 'core-components.js')),
     relative(path.join(Assemble.paths.plugins, 'core-pagination-step2.js'))
   ],
-  // Core middleware to be loaded by dovetail
-  middleware: [
-    middleware.config,
-    middleware.data,
-    middleware.build,
-    middleware.render
+  // Core steps
+  core: [
+    core.config,
+    core.data,
+    core.build,
+    core.render
   ]
 }, file.readDataSync(Assemble.paths.defaults));
-
-Assemble.prototype.use = function(middleware) {
-  this.middleware.push(middleware);
-  return this;
-};
 
 Assemble.prototype.build = function (callback) {
 
   // reset the context before building
   this.context = this.utils.context.stub(this);
 
-  // loop over each middleware
+  // loop over each core step
   async.eachSeries(
 
-    (this.middleware.length === 0 ? Assemble.defaults.middleware : this.middleware),
+    Assemble.defaults.core,
 
-    // call each middleware
-    function (middleware, next) {
-      middleware(this, next);
+    // call each step
+    function (step, next) {
+      step(this, next);
     }.bind(this),
 
     // call done with the callback
-    middleware.done(this, callback));
+    core.done(this, callback));
 
 };
