+++ lib/core/data.js
