+++ lib/core/index.js
