+++ lib/core/pages.js
