+++ lib/core/partials.js
