+++ lib/core/render.js
