+++ lib/engines/handlebars/handlebars.js
@@ -14,7 +14,7 @@ var async = require('async');
 var _ = require('lodash');
 
 // local
-var notifier = require('../../middleware/notifier');
+var notifier = require('../../core/notifier');
 var componentHandler = require('./components');
 var asyncHelpers = require('./async');
 var layouts = require('./layouts');
